add more log

+++ source/dnode/vnode/src/inc/tqInt.h
@@ -206,7 +206,7 @@ typedef struct {
 
 typedef struct {
   int64_t consumerId;
-  int64_t epoch;
+  int32_t epoch;
   char    cgroup[TSDB_TOPIC_FNAME_LEN];
   SArray* topics;  // SArray<STqTopic>
 } STqConsumer;
